rag-app_requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv
chromadb
tqdm
pdfplumber
python-docx
openai
requests
pytest
pytest-asyncio
httpx
nltk
python-dateutil
langchain>=0.0.420
pinecone-client
weaviate-client
